requirements.txt
Django==4.1.3
psycopg2-binary==2.9.5
python-dotenv==0.21.0
whitenoise==6.2.0
gunicorn==20.1.0